+++ cmake/Dependencies.cmake
@@ -108,3 +108,32 @@ macro (get_pc_package vala_package pc_package )
     #    set( pc_package ${vala_package} )
     #endif()
 endmacro()
+
+macro(add_local_package vala_package dependencies)
+    list (APPEND list_vala_local_packages ${vala_package})
+    string (STRIP ${dependencies} deps)
+    list (APPEND list_vala_local_packages_deps ${deps})
+endmacro()
+
+# Not used!!!
+macro (is_vala_package_local package_name is_local)
+     if( package_name STREQUAL "vala-stacktrace")
+        set (is_local "true")
+    else ()
+        set (is_local "false")
+    endif()
+endmacro()
+
+macro (local_check_package vala_package)
+
+    list(FIND list_vala_local_packages ${vala_package} index)
+
+    if( index GREATER -1 )
+        list(GET list_vala_local_packages_deps ${index} deps)
+        set (DEPS_LIBRARIES ${DEPS_LIBRARIES} "-l${vala_package}")
+        set(VALA_PACKAGES ${VALA_PACKAGES} "${deps}")
+    else ()
+        message ( FATAL_ERROR "${FatalColor}Local vala package not found${NC}: ${vala_package}") 
+    endif()
+
+endmacro()