+++ cmake/tasks/ElementaryApp.cmake
@@ -1,3 +1,9 @@
+# Bundle version: 0.9
+#
+# File History:
+#    - 0.1 : refactoring
+#    - 0.2 : fix library with depending app
+
 macro(build_elementary_app)
     parse_arguments(ARGS "BINARY_NAME;TITLE;VERSION;RELEASE_NAME;SOURCE_PATH;VALA_FILES;C_FILES;VALA_DEFINES;PACKAGES;C_DEFINES;SCHEMA;VALA_OPTIONS;C_OPTIONS;ICON;DESKTOP" "" ${ARGN})
 
@@ -28,7 +34,7 @@ macro(build_elementary_app)
         message( FATAL_ERROR "${FatalColor}The DESKTOP file doesn't exist${NC}. File: ${CMAKE_CURRENT_SOURCE_DIR}/${ARGS_DESKTOP}")
     endif()
 
-    prepare_elementary (
+    hen_build (
         BINARY_NAME
             ${ARGS_BINARY_NAME}
         TITLE
@@ -60,6 +66,9 @@ macro(build_elementary_app)
     )
 
     add_executable (${ARGS_BINARY_NAME} ${VALA_C} ${C_FILES})
+    foreach( vala_local_pkg ${list_vala_local_packages})
+        add_dependencies (${ARGS_BINARY_NAME}  "${vala_local_pkg}")
+    endforeach()
     target_link_libraries (${ARGS_BINARY_NAME} ${DEPS_LIBRARIES})
 
     install_elementary_app (${ARGS_BINARY_NAME} ${ARGS_ICON} "${ARGS_DESKTOP}")