+++ cmake/tasks/ElementaryCli.cmake
@@ -1,3 +1,8 @@
+# Bundle version: 0.9
+#
+# File History:
+#    - 0.1 : refactoring
+#    - 0.2 : fix library with depending app
 
 macro(build_elementary_cli)
     parse_arguments(ARGS "BINARY_NAME;TITLE;VERSION;RELEASE_NAME;SOURCE_PATH;VALA_FILES;C_FILES;VALA_DEFINES;PACKAGES;C_DEFINES;SCHEMA;VALA_OPTIONS;C_OPTIONS" "" ${ARGN})
@@ -6,7 +11,7 @@ macro(build_elementary_cli)
     set (PKGDATADIR "")
     set (GETTEXT_PACKAGE "${ARGS_BINARY_NAME}")
 
-    prepare_elementary (
+    hen_build (
         BINARY_NAME
             ${ARGS_BINARY_NAME}
         TITLE
@@ -38,6 +43,9 @@ macro(build_elementary_cli)
     )
 
     add_executable (${ARGS_BINARY_NAME} ${VALA_C} ${C_FILES})
+    foreach( vala_local_pkg ${list_vala_local_packages})
+        add_dependencies (${ARGS_BINARY_NAME}  "${vala_local_pkg}")
+    endforeach()
     target_link_libraries (${ARGS_BINARY_NAME} ${DEPS_LIBRARIES})
 
     install_elementary_cli (${ARGS_BINARY_NAME})