+++ cmake/tasks/ElementaryLibrary.cmake
@@ -1,3 +1,9 @@
+# Bundle version: 0.9
+#
+# File History:
+#    - 0.1 : refactoring
+#    - 0.2 : fix glib.h: No such file or directory (vala-stacktrace working)
+#    - 0.3 : fix library with depending app
 
 macro(build_elementary_library)
     parse_arguments(ARGS "BINARY_NAME;TITLE;VERSION;RELEASE_NAME;SOVERSION;LINKING;SOURCE_PATH;VALA_FILES;C_FILES;VALA_DEFINES;PACKAGES;C_DEFINES;SCHEMA;VALA_OPTIONS;C_OPTIONS" "" ${ARGN})
@@ -14,7 +20,7 @@ macro(build_elementary_library)
     set (PKGDATADIR "")
     set (GETTEXT_PACKAGE "${ARGS_BINARY_NAME}")
 
-    prepare_elementary (
+    hen_build (
         BINARY_NAME
             ${ARGS_BINARY_NAME}
         TITLE
@@ -52,6 +58,7 @@ macro(build_elementary_library)
     # TODO fix the output path
     configure_file (${DIR_ELEMENTARY_TEMPLATES}/lib.pc.cmake ${CMAKE_BINARY_DIR}/${ARGS_BINARY_NAME}.pc)
     configure_file (${DIR_ELEMENTARY_TEMPLATES}/lib.deps.cmake ${CMAKE_BINARY_DIR}/${ARGS_BINARY_NAME}.deps)
+    add_local_package (${ARGS_BINARY_NAME} ${COMPLETE_DIST_PC_PACKAGES} )
 
     if( ARGS_LINKING STREQUAL "static")
         add_library (${ARGS_BINARY_NAME} STATIC ${VALA_C} ${C_FILES})
@@ -68,7 +75,6 @@ macro(build_elementary_library)
             SOVERSION ${ARGS_SOVERSION}
         )
     endif()
-
     target_link_libraries (${ARGS_BINARY_NAME} ${DEPS_LIBRARIES})
 
     install_elementary_library (${ARGS_BINARY_NAME} ${ARGS_LINKING})