+++ cmake/tasks/ElementaryPlug.cmake
@@ -1,3 +1,9 @@
+# Bundle version: 0.9
+#
+# File History:
+#    - 0.1 : refactoring
+#    - 0.2 : fix library with depending app
+
 macro(build_elementary_plug)
     parse_arguments(ARGS "BINARY_NAME;TITLE;VERSION;RELEASE_NAME;PLUG_CATEGORY;SOURCE_PATH;VALA_FILES;C_FILES;PACKAGES;C_DEFINES;VALA_DEFINES;SCHEMA;VALA_OPTIONS;C_OPTIONS" "" ${ARGN})
 
@@ -9,7 +15,7 @@ macro(build_elementary_plug)
     set (PKGDATADIR "${DATADIR}/${ARGS_PLUG_CATEGORY}/${ARGS_BINARY_NAME}")
     set (GETTEXT_PACKAGE "${ARGS_BINARY_NAME}")
 
-    prepare_elementary (
+    hen_build (
         BINARY_NAME
             ${ARGS_BINARY_NAME}
         TITLE
@@ -44,6 +50,9 @@ macro(build_elementary_plug)
     )
 
     add_library (${ARGS_BINARY_NAME} MODULE ${VALA_C} ${C_FILES})
+    foreach( vala_local_pkg ${list_vala_local_packages})
+        add_dependencies (${ARGS_BINARY_NAME}  "${vala_local_pkg}")
+    endforeach()
     target_link_libraries (${ARGS_BINARY_NAME} ${DEPS_LIBRARIES})
 
     install_elementary_plug (${ARGS_BINARY_NAME})