+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+* 0.9.4
+ - add list options
+ - add colors
+ - fix library task (vala-stacktrace working)
+ - add --target-glib 2.32 by default (and emit a warning if set in the cmake file)
+
 * 0.9.3
-  - add `hen.sh` that provides update feature
-  - refactoring
+ - add `hen.sh` that provides update feature
+ - refactoring (eidete working)
+ - add build and prepare options
   
 * Initial revision: 0.9